Website/requirements.txt
botocore==1.4.92
decorator==4.0.10
gunicorn==19.6.0
Django==1.10.6
djangorestframework==3.6.3
django-filter==1.0.4
docutils==0.13.1
httpie==0.9.9
Pillow==3.4.2
psycopg2==2.7.1
requests==2.9.1
whitenoise==3.3.0